moving memory.rs out of datafusion/core
