+++ datafusion/catalog/Cargo.toml
@@ -35,6 +35,9 @@ datafusion-execution = { workspace = true }
 datafusion-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
 parking_lot = { workspace = true }
+dashmap = { workspace = true }
+itertools = { workspace = true }
+
 
 [dev-dependencies]
 tokio = { workspace = true }
