+++ datafusion/core/src/catalog_common/mod.rs
@@ -24,12 +24,8 @@
 
 pub mod information_schema;
 pub mod listing_schema;
-pub mod memory;
 
 pub use crate::catalog::{CatalogProvider, CatalogProviderList, SchemaProvider};
-pub use memory::{
-    MemoryCatalogProvider, MemoryCatalogProviderList, MemorySchemaProvider,
-};
 
 pub use datafusion_sql::{ResolvedTableReference, TableReference};
