+++ datafusion/core/tests/catalog/mod.rs
@@ -1,0 +1,1 @@
+mod memory;
